fix the relative import of shared/utils type definitions in shared/types route code
