+++ shared/utils/devTs.ts
@@ -1,3 +1,4 @@
 // please list in alphanumeric order for readability
 // use .ts file extension for all code files below
 export * from './constants/AiHisto.ts'
+export * from './constants/terms.ts'
